+++ examples/live/binance_futures_testnet_market_maker.py
@@ -21,7 +21,7 @@ from nautilus_trader.adapters.binance.config import BinanceDataClientConfig
 from nautilus_trader.adapters.binance.config import BinanceExecClientConfig
 from nautilus_trader.adapters.binance.factories import BinanceLiveDataClientFactory
 from nautilus_trader.adapters.binance.factories import BinanceLiveExecClientFactory
-from nautilus_trader.config import CacheDatabaseConfig
+from nautilus_trader.config import CacheConfig
 from nautilus_trader.config import InstrumentProviderConfig
 from nautilus_trader.config import LiveExecEngineConfig
 from nautilus_trader.config import LoggingConfig
@@ -52,8 +52,8 @@ config_node = TradingNodeConfig(
         reconciliation_lookback_mins=1440,
         filter_position_reports=True,
     ),
-    cache_database=CacheDatabaseConfig(
-        type="in-memory",
+    cache=CacheConfig(
+        # database=DatabaseConfig(),
         flush_on_start=False,
         timestamps_as_iso8601=True,
     ),