+++ nautilus_trader/config/common.py
@@ -108,23 +108,6 @@ class NautilusConfig(msgspec.Struct, kw_only=True, frozen=True):
         return bool(msgspec.json.decode(self.json(), type=self.__class__))
 
 
-class CacheConfig(NautilusConfig, frozen=True):
-    """
-    Configuration for ``Cache`` instances.
-
-    Parameters
-    ----------
-    tick_capacity : PositiveInt, default 10_000
-        The maximum length for internal tick dequeues.
-    bar_capacity : PositiveInt, default 10_000
-        The maximum length for internal bar dequeues.
-
-    """
-
-    tick_capacity: PositiveInt = 10_000
-    bar_capacity: PositiveInt = 10_000
-
-
 class DatabaseConfig(NautilusConfig, frozen=True):
     """
     Configuration for database connections.
@@ -146,7 +129,7 @@ class DatabaseConfig(NautilusConfig, frozen=True):
 
     Notes
     -----
-    Requires Redis version 6.2.0 and above for correct operation.
+    If `type` is 'redis' then requires Redis version 6.2.0 and above for correct operation.
 
     """
 
@@ -158,26 +141,19 @@ class DatabaseConfig(NautilusConfig, frozen=True):
     ssl: bool = False
 
 
-class CacheDatabaseConfig(NautilusConfig, frozen=True):
+class CacheConfig(NautilusConfig, frozen=True):
     """
-    Configuration for ``CacheDatabase`` instances.
+    Configuration for ``Cache`` instances.
 
     Parameters
     ----------
-    type : str, {'in-memory', 'redis'}, default 'in-memory'
-        The database type.
-    host : str, default 'localhost'
-        The database host address.
-    port : int, optional
-        The database port.
-    username : str, optional
-        The account username for the database connection.
-    password : str, optional
-        The account password for the database connection.
-    ssl : bool, default False
-        If database should use an SSL enabled connection.
+    database : DatabaseConfig, optional
+        The configuration for the cache backing database.
     encoding : str, {'msgpack', 'json'}, default 'msgpack'
         The encoding for database operations, controls the type of serializer used.
+    timestamps_as_iso8601, default False
+        If timestamps should be persisted as ISO 8601 strings.
+        If `False` then will persit as UNIX nanoseconds.
     buffer_interval_ms : PositiveInt, optional
         The buffer interval (milliseconds) between pipelined/batched transactions.
         The recommended range if using buffered pipeling is [10, 1000] milliseconds,
@@ -188,24 +164,22 @@ class CacheDatabaseConfig(NautilusConfig, frozen=True):
         If a 'trader-' prefix is applied to keys.
     use_instance_id : bool, default False
         If the traders instance ID should be used for keys.
-    timestamps_as_iso8601, default False
-        If timestamps should be persisted as ISO 8601 strings.
-        If `False` then will persit as UNIX nanoseconds.
+    tick_capacity : PositiveInt, default 10_000
+        The maximum length for internal tick dequeues.
+    bar_capacity : PositiveInt, default 10_000
+        The maximum length for internal bar dequeues.
 
     """
 
-    type: str = "in-memory"
-    host: str = "localhost"
-    port: int | None = None
-    username: str | None = None
-    password: str | None = None
-    ssl: bool = False
+    database: DatabaseConfig | None = None
     encoding: str = "msgpack"
+    timestamps_as_iso8601: bool = False
     buffer_interval_ms: PositiveInt | None = None
     flush_on_start: bool = False
     use_trader_prefix: bool = True
     use_instance_id: bool = False
-    timestamps_as_iso8601: bool = False
+    tick_capacity: PositiveInt = 10_000
+    bar_capacity: PositiveInt = 10_000
 
 
 class MessageBusConfig(NautilusConfig, frozen=True):
@@ -218,6 +192,9 @@ class MessageBusConfig(NautilusConfig, frozen=True):
         The configuration for the message bus backing database.
     encoding : str, {'msgpack', 'json'}, default 'msgpack'
         The encoding for database operations, controls the type of serializer used.
+    timestamps_as_iso8601, default False
+        If timestamps should be persisted as ISO 8601 strings.
+        If `False` then will persit as UNIX nanoseconds.
     buffer_interval_ms : PositiveInt, optional
         The buffer interval (milliseconds) between pipelined/batched transactions.
         The recommended range if using buffered pipeling is [10, 1000] milliseconds,
@@ -232,9 +209,6 @@ class MessageBusConfig(NautilusConfig, frozen=True):
         The additional prefix for externally published stream names (must have a `database` config).
     use_instance_id : bool, default False
         If the traders instance ID should be used in stream names.
-    timestamps_as_iso8601, default False
-        If timestamps should be persisted as ISO 8601 strings.
-        If `False` then will persit as UNIX nanoseconds.
     types_filter : list[type], optional
         A list of serializable types *not* to publish externally.
 
@@ -242,11 +216,11 @@ class MessageBusConfig(NautilusConfig, frozen=True):
 
     database: DatabaseConfig | None = None
     encoding: str = "msgpack"
+    timestamps_as_iso8601: bool = False
     buffer_interval_ms: PositiveInt | None = None
     autotrim_mins: int | None = None
     stream: str | None = None
     use_instance_id: bool = False
-    timestamps_as_iso8601: bool = False
     types_filter: list[type] | None = None
 
 
@@ -783,8 +757,6 @@ class NautilusKernelConfig(NautilusConfig, frozen=True):
         The trader ID for the kernel (must be a name and ID tag separated by a hyphen).
     cache : CacheConfig, optional
         The cache configuration.
-    cache_database : CacheDatabaseConfig, optional
-        The cache database configuration.
     message_bus : MessageBusConfig, optional
         The message bus configuration.
     data_engine : DataEngineConfig, optional
@@ -847,7 +819,6 @@ class NautilusKernelConfig(NautilusConfig, frozen=True):
     trader_id: str
     instance_id: str | None = None
     cache: CacheConfig | None = None
-    cache_database: CacheDatabaseConfig | None = None
     message_bus: MessageBusConfig | None = None
     data_engine: DataEngineConfig | None = None
     risk_engine: RiskEngineConfig | None = None