+++ RELEASES.md
@@ -8,8 +8,10 @@ Released on TBD (UTC).
 - Removed `redis` and `hiredis` dependencies from Python codebase
 
 ### Breaking Changes
+- Changed `CacheConfig` to take a `DatabaseConfig` (better symmetry with `MessageBusConfig`)
 - Changed `RedisCacheDatabase` data structure for currencies from hashset to simpler key-value (you'll need to clear cache or delete all curreny keys)
 - Changed `Actor` state loading to now use the standard `Serializer`
+- Removed `CacheDatabaseConfig` (due above config change)
 - Removed `infrastructure` subpackage (now redundant with new Rust implementation)
 
 ### Fixes